move things and add a script
