+++ package.json
@@ -21,8 +21,10 @@
         "dev": "npm run dev --workspaces",
         "dev:tauri": "nix develop .#tauri-shell -c cargo tauri dev --verbose",
         "test": "npm run test --workspaces",
-        "lint:all": "npm run lint --workspaces --if-present && npm run format:check --workspaces --if-present",
-        "format:all": "npm run format --workspaces --if-present"
+        "lint:all": "npm run lint --workspaces --if-present",
+        "format:all": "npm run format --workspaces --if-present",
+        "check:all": "npm run check --workspaces --if-present",
+        "lint-format-check:all": "npm run lint:all && npm run format:all && npm run check:all" 
     },
     "devDependencies": {
         "@tanstack/svelte-query": "^5.59.20",
