+++ packages/ui-components/src/__tests__/CardProperty.test.ts
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/svelte';
 import { describe, it, expect } from 'vitest';
-import CardProperty from './CardProperty.svelte';
+import CardProperty from '../lib/components/CardProperty.svelte';
 
 describe('Page Component', () => {
 	it('should render the card component', async () => {
