+++ tauri-app/src/lib/components/tables/VaultListTable.svelte
@@ -95,6 +95,7 @@
         </div>
         <div class="flex flex-col items-end gap-y-2">
           <ListViewOrderbookFilters
+            {query}
             {activeSubgraphs}
             {settings}
             {accounts}
